feature: get magnitude of images & add setup.py of repository
Before getting the magnitude, files with median zero are discarded. Since each image comes from the difference of two consecutive observations, before the magnitude the absolute value is computed. the magnitude is given by: -2.5*np.log10(image). To mitigate noise, magnitudes larger than the median of the magnitude are replaced by the median.